requirements.txt
certifi==2018.4.16
chardet==3.0.4
dj-database-url==0.5.0
Django==1.11.12
django-appconf==1.0.2
django-autoslug==1.9.3
django-bootstrap-form==3.4
django-filter==1.1.0
django-rest-framework==0.1.0
django-widget-tweaks==1.4.2
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
idna==2.6
Markdown==2.6.11
Pillow==5.1.0
PyJWT==1.6.3
python-decouple==3.1
pytz==2018.4
requests==2.18.4
urllib3==1.22
